// src/rest/model/fileType.ts
/**
 * Pakkasmarja REST API
 * REST API for Pakkasmarja
 *
 * OpenAPI spec version: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export type FileType = 'IMAGE' | 'PDF' | 'OTHER' | 'FOLDER';

export const FileType = {
    IMAGE: 'IMAGE' as FileType,
    PDF: 'PDF' as FileType,
    OTHER: 'OTHER' as FileType,
    FOLDER: 'FOLDER' as FileType
};